+++ library-api/src/test/java/com/marquinhus/libraryapi/service/BookServiceTest.java
@@ -39,7 +39,7 @@ public class BookServiceTest {
         Mockito.when(repository.existsByIsbn(Mockito.anyString())).thenReturn(false);
         Mockito.when(repository.save(book)).thenReturn(
                 Book.builder()
-                        .id(11)
+                        .id((long) 11)
                         .isbn("123")
                         .title("As aventuras")
                         .author("Fulano")
